requirements.txt
Django==1.5.1
wsgiref==0.1.2
south==0.7.6
redis==2.7.2
celery==3.0.12
django-celery==3.0.11
dj-database-url==0.2.1
psycopg2==2.4.5
gunicorn==0.17.2
django-heroku-memcacheify==0.4
pylibmc==1.2.2
django-pylibmc-sasl==0.2.4
markdown==2.3.1
google-api-python-client==1.1
gdata==2.0.17
pil==1.1.7
django-tinymce==1.5.1